Added support for stale-while-revalidate of the HTML edge cache

+++ streaming-optimizations/streaming-optimizations.js
@@ -10,6 +10,15 @@ const CLOUDFLARE_API = {
   zone: ""   // "Zone ID" from the API section of the dashboard overview page https://dash.cloudflare.com/
 };
 
+// Default cookie prefixes for bypass when HTML edge cache is enabled
+const DEFAULT_BYPASS_COOKIES = [
+  "wp-",
+  "wordpress",
+  "comment_",
+  "woocommerce_"
+];
+
+
 /******************************************************************************
  *  Main control flow
  *****************************************************************************/
@@ -191,7 +200,7 @@ async function processRequest(request, event) {
   * @param {Request} request - Original request
   * @returns {bool} true if edge caching is enabled
   */
- function isEdgeCacheEnabled(request) {
+function isEdgeCacheEnabled(request) {
   let enabled = ENABLE_EDGE_CACHE;
 
   // Disable if there is a cache in front of us
@@ -216,6 +225,9 @@ async function processRequest(request, event) {
  * @param {Event} event - original event
  */
 async function edgeCacheFetch(originalRequest, event) {
+  let cfCacheStatus = null;
+  const accept = originalRequest.headers.get('Accept');
+  const isHTML = (accept && accept.indexOf('text/html') >= 0);
   let {response, cacheVer, status, bypassCache} = await getEdgeCachedResponse(originalRequest);
 
   if (response === null) {
@@ -226,23 +238,42 @@ async function edgeCacheFetch(originalRequest, event) {
 
     if (response) {
       const options = getEdgeCacheResponseOptions(response);
-      if (options.purge) {
+      if (options && options.purge) {
         await purgeEdgeCache(cacheVer, event);
         status += ', Purged';
       }
       bypassCache = bypassCache || shouldBypassEdgeCache(request, response);
-      if (options.cache && !bypassCache) {
+      if ((!options || options.cache) && isHTML &&
+          originalRequest.method === 'GET' && response.status === 200 &&
+          !bypassCache) {
         status += await cacheEdgeResponse(cacheVer, originalRequest, response, event);
+      }
+    }
+  } else {
+    // If the origin didn't send the control header we will send the cached response but update
+    // the cached copy asynchronously (stale-while-revalidate). This commonly happens with
+    // a server-side disk cache that serves the HTML directly from disk.
+    cfCacheStatus = 'HIT';
+    if (originalRequest.method === 'GET' && response.status === 200 && isHTML) {
+      bypassCache = bypassCache || shouldBypassEdgeCache(originalRequest, response);
+      if (!bypassCache) {
+        const options = getEdgeCacheResponseOptions(response);
+        if (!options) {
+          status += ', Refreshed';
+          event.waitUntil(updateEdgeCache(originalRequest, cacheVer, event));
+        }
       }
     }
   }
 
-  const accept = originalRequest.headers.get('Accept');
-  if (response && status !== null && originalRequest.method === 'GET' && response.status === 200 && accept && accept.indexOf('text/html') >= 0) {
+  if (response && status !== null && originalRequest.method === 'GET' && response.status === 200 && isHTML) {
     response = new Response(response.body, response);
     response.headers.set('x-HTML-Edge-Cache-Status', status);
     if (cacheVer !== null) {
       response.headers.set('x-HTML-Edge-Cache-Version', cacheVer.toString());
+    }
+    if (cfCacheStatus) {
+      response.headers.set('CF-Cache-Status', cfCacheStatus);
     }
   }
 
@@ -264,11 +295,15 @@ function shouldBypassEdgeCache(request, response) {
   if (request && response) {
     const options = getEdgeCacheResponseOptions(response);
     const cookieHeader = request.headers.get('cookie');
-    if (cookieHeader && cookieHeader.length && options.bypassCookies.length) {
+    let bypassCookies = DEFAULT_BYPASS_COOKIES;
+    if (options) {
+      bypassCookies = options.bypassCookies;
+    }
+    if (cookieHeader && cookieHeader.length && bypassCookies.length) {
       const cookies = cookieHeader.split(';');
       for (let cookie of cookies) {
         // See if the cookie starts with any of the logged-in user prefixes
-        for (let prefix of options.bypassCookies) {
+        for (let prefix of bypassCookies) {
           if (cookie.trim().startsWith(prefix)) {
             bypassCache = true;
             break;
@@ -327,16 +362,16 @@ async function getEdgeCachedResponse(request) {
           status = 'Bypass Cookie';
         } else {
           status = 'Hit';
-          response = cachedResponse;
-          response.headers.delete('x-HTML-Edge-Cache');
-          response.headers.delete('Cache-Control');
+          cachedResponse.headers.delete('Cache-Control');
+          cachedResponse.headers.delete('x-HTML-Edge-Cache-Status');
           for (header of CACHE_HEADERS) {
-            let value = response.headers.get('x-HTML-Edge-Cache-' + header);
+            let value = cachedResponse.headers.get('x-HTML-Edge-Cache-Header-' + header);
             if (value) {
-              response.headers.delete('x-HTML-Edge-Cache-' + header);
-              response.headers.set(header, value);
+              cachedResponse.headers.delete('x-HTML-Edge-Cache-Header-' + header);
+              cachedResponse.headers.set(header, value);
             }
           }
+          response = cachedResponse;
         }
       } else {
         status = 'Miss';
@@ -375,6 +410,31 @@ async function purgeEdgeCache(cacheVer, event) {
 }
 
 /**
+ * Update the cached copy of the given page
+ * @param {Request} originalRequest - Original Request
+ * @param {String} cacheVer - Cache Version
+ * @param {EVent} event - Original event
+ */
+async function updateEdgeCache(originalRequest, cacheVer, event) {
+  // Clone the request, add the edge-cache header and send it through.
+  let request = new Request(originalRequest);
+  request.headers.set('x-HTML-Edge-Cache', 'supports=cache|purgeall|bypass-cookies');
+  response = await fetch(request);
+
+  if (response) {
+    status = ': Fetched';
+    const options = getEdgeCacheResponseOptions(response);
+    if (options && options.purge) {
+      await purgeCache(cacheVer, event);
+    }
+    let bypassCache = shouldBypassEdgeCache(request, response);
+    if ((!options || options.cache) && !bypassCache) {
+      await cacheEdgeResponse(cacheVer, originalRequest, response, event);
+    }
+  }
+}
+
+/**
  * Cache the returned content (but only if it was a successful GET request)
  * 
  * @param {Int} cacheVer - Current cache version (if already retrieved)
@@ -401,7 +461,7 @@ async function cacheEdgeResponse(cacheVer, request, originalResponse, event) {
         let value = response.headers.get(header);
         if (value) {
           response.headers.delete(header);
-          response.headers.set('x-HTML-Edge-Cache-' + header, value);
+          response.headers.set('x-HTML-Edge-Cache-Header-' + header, value);
         }
       }
       response.headers.delete('Set-Cookie');
@@ -409,8 +469,7 @@ async function cacheEdgeResponse(cacheVer, request, originalResponse, event) {
       event.waitUntil(cache.put(cacheKeyRequest, response));
       status = ", Cached";
     } catch (err) {
-      // Send the exception back in the response header for debugging
-      status = ", Cache Write Exception: " + err.message;
+      // status = ", Cache Write Exception: " + err.message;
     }
   }
   return status;
@@ -422,14 +481,14 @@ async function cacheEdgeResponse(cacheVer, request, originalResponse, event) {
  * @returns {*} Parsed commands
  */
 function getEdgeCacheResponseOptions(response) {
-  let options = {
-    purge: false,
-    cache: false,
-    bypassCookies: []
-  };
-
+  let options = null;
   let header = response.headers.get('x-HTML-Edge-Cache');
   if (header) {
+    options = {
+      purge: false,
+      cache: false,
+      bypassCookies: []
+    };
     let commands = header.split(',');
     for (let command of commands) {
       if (command.trim() === 'purgeall') {
